requirements.txt
langchain-core
langchain-openai
langgraph
python-dotenv
langsmith
fastapi
uvicorn
coloredlogs
tiktoken
langchain-text-splitters
langchain_chroma